shellcheck: SC2046
Previously, we were redirecting stderr to where stdout was currently pointing to
and then redirecting stdout to /dev/null. As a result, stdout would be hidden,
but stderr would be displayed on the terminal. I don't think that was intentional.

This changes the script so both stdout and stderr are redirected to /dev/null. If `grep`
exits with non-zero, then we print the entire `tmp_stderr` file containing the
stderr from the curl, wget, etc command.