Update rocky Dockerfiles
Add README describing how to run these Dockerfiles

* Dockerfile is used to verify upgrades, by default from 7.34.0 to 8.10.0.
* Dockerfile.versions is used to install multiple release packages and verify
  you can get a listing.

Update `Dockerfile` and `Dockerfile.versions` to allow credentials to be
specified, taking care not to save the credentials in the docker image.

Update `Dockerfile.versions` to download the puppet8 release package from
yum-support and after installation, configure the username and password in the
repo config.

Added `run-versions.shUpdate `Dockerfile` to copy `run-upgrade.sh` into the
docker image and run it using an externally provided password.

Update `run-upgrade.sh` to pass username and password as facts to `upgrade.pp`.
When the upgrade is performed, the module will create pc_repo with the provided
credentials and use that to upgrade to 8.10.0.